requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
fintekkers-ledger-models==0.1.91
grpcio==1.53.0
idna==3.4
jupyter==1.0.0
kaleido==0.2.1
numpy==1.24.2
pandas==2.0.0
plotly==5.23.0
protobuf==4.22.1
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
tzdata==2023.3
urllib3==1.26.15
psycopg2-binary==2.9.11
yahoofinancials==1.20

